refactor optionListEditor to make it more readable
