+++ frontend/packages/ux-editor/src/components/config/editModal/EditOptions/OptionTabs/EditTab/OptionListEditor/ManualOptionsEditor/ManualOptionsEditor.module.tsx.css
@@ -1,0 +1,11 @@
+.editOptionTabModal[open] {
+  --code-list-modal-min-width: min(80rem, 100%);
+  --code-list-modal-height: min(45rem, 100%);
+
+  min-width: var(--code-list-modal-min-width);
+  height: var(--code-list-modal-height);
+}
+
+.content {
+  height: 100%;
+}
